# ARROW-7166: [Java] Remove redundant code for Jdbc adapters