As discussed in https://github.com/apache/arrow/pull/5508#issuecomment-543011016, we need a separate issue to extract common logic to a common super class.

This makes the code clearer, and we need to make sure we have no performance regression.