@liyafan82 2019-11-19
Benchmark evaluations show no performance degradation:

Before:
JdbcAdapterBenchmarks.consumeBenchmark      avgt    5   547.267 ±   8.121  us/op
JdbcAdapterBenchmarks.consumeRowsBenchmark  avgt    5   593.117 ± 109.100  us/op
JdbcAdapterBenchmarks.testJdbcToArrow       avgt    5  5786.141 ±  68.501  us/op

After: 
JdbcAdapterBenchmarks.consumeBenchmark      avgt    5   554.501 ±  17.261  us/op
JdbcAdapterBenchmarks.consumeRowsBenchmark  avgt    5   558.829 ±  22.413  us/op
JdbcAdapterBenchmarks.testJdbcToArrow       avgt    5  5757.178 ± 197.539  us/op